refactor http queries for account details; - use a single http client to make use of HTTP keepalive - move any remaining structs from api module into the model module
