+++ client/cosmos/api/account.go
@@ -2,42 +2,24 @@ package api
 
 import (
 	"encoding/json"
-	"fmt"
-	"io"
 	"net/http"
 
+	"github.com/informalsystems/stakooler/client/cosmos"
 	"github.com/informalsystems/stakooler/client/cosmos/model"
 )
 
-func GetAuth(account *model.Account) (model.AuthResponse, error) {
-	var authResponse model.AuthResponse
+func GetAuth(account *model.Account, client *http.Client) (response model.AuthResponse, err error) {
+	var body []byte
 
 	url := account.Chain.LCD + "/cosmos/auth/v1beta1/accounts/" + account.Address
-	method := "GET"
+	body, err = cosmos.HttpGet(url, client)
+	if err != nil {
+		return
+	}
 
-	client := &http.Client{}
-	req, err := http.NewRequest(method, url, nil)
-
+	err = json.Unmarshal(body, &response)
 	if err != nil {
-		fmt.Println(err)
-		return authResponse, err
+		return
 	}
-	res, err := client.Do(req)
-	if err != nil {
-		fmt.Println(err)
-		return authResponse, err
-	}
-	defer res.Body.Close()
-
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		fmt.Println(err)
-		return authResponse, err
-	}
-	err = json.Unmarshal(body, &authResponse)
-	if err != nil {
-		fmt.Println(err)
-		return authResponse, err
-	}
-	return authResponse, nil
+	return
 }
