+++ client/cosmos/model/distribution.go
@@ -1,0 +1,24 @@
+package model
+
+type RewardsResponse struct {
+	Rewards []struct {
+		ValidatorAddress string `json:"validator_address"`
+		Reward           []struct {
+			Denom  string `json:"denom"`
+			Amount string `json:"amount"`
+		} `json:"reward"`
+	} `json:"rewards"`
+	Total []struct {
+		Denom  string `json:"denom"`
+		Amount string `json:"amount"`
+	} `json:"total"`
+}
+
+type CommissionResponse struct {
+	Commissions struct {
+		Commission []struct {
+			Denom  string `json:"denom"`
+			Amount string `json:"amount"`
+		} `json:"commission"`
+	} `json:"commission"`
+}
